+++ examples/x6-app-dag/src/pages/common/hooks/useSafeSetHtml.ts
