+++ examples/x6-app-draw/.umirc.ts
@@ -2,9 +2,6 @@ import { defineConfig } from 'umi'
 
 export default defineConfig({
   publicPath: './',
-  nodeModulesTransform: {
-    type: 'none',
-  },
   routes: [
     { path: '/', component: '@/pages/index' },
     { path: '/apps/draw', component: '@/pages/index' },
