+++ examples/x6-app-er/.umirc.ts
@@ -2,9 +2,6 @@ import { defineConfig } from 'umi';
 
 export default defineConfig({
   publicPath: './',
-  nodeModulesTransform: {
-    type: 'none',
-  },
   routes: [
     { path: '/', component: '@/pages/index' },
     { path: '/apps/er', component: '@/pages/index' },
